requirements.txt
tuya-iot-py-sdk
python-dotenv
colorama